requirements.txt
dnspython==1.16.0
colorama==0.3.9
